requirements.txt
# ArgusAI — Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
tensorflow>=2.13.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.0.0

# Persistence
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0